RTS/DTR for sending config
A programming device with just RTS to reset and DTR to GPIO0 works with esptool but does not then work for sending config. This is because tasmotizer leaves RTS/DTR to QSerialPort which sets active (low) even when no flow control (default). By setting inactive (high) after opening the port, this therefore causes a clean reset in the same way as esptool does. The sleep allows time for tasmota to start and be ready to accept the config. Perhaps this could be based on the "self-resetting" option, or be a separate option in its own right. I believe it should work with devices like esptool does though.